+++ src/transport/batchedNetworkInterface.ts
@@ -2,8 +2,6 @@ import {
   ExecutionResult,
 } from 'graphql';
 
-import 'whatwg-fetch';
-
 import {
   HTTPFetchNetworkInterface,
   HTTPNetworkInterface,