+++ test/ObservableQuery.ts
@@ -215,7 +215,7 @@ describe('ObservableQuery', () => {
       const variables2 = { first: 1 };
 
 
-      const observable: ObservableQuery = mockWatchQuery({
+      const observable: ObservableQuery<any> = mockWatchQuery({
         request: { query: queryWithVars, variables: variables1 },
         result: { data },
       }, {
@@ -236,7 +236,7 @@ describe('ObservableQuery', () => {
     });
 
     it('does a network request if forceFetch becomes true', (done) => {
-      const observable: ObservableQuery = mockWatchQuery({
+      const observable: ObservableQuery<any> = mockWatchQuery({
         request: { query, variables },
         result: { data: dataOne },
       }, {
@@ -257,7 +257,7 @@ describe('ObservableQuery', () => {
 
     it('does a network request if noFetch becomes true then store is reset then noFetch becomes false', (done) => {
       let queryManager: QueryManager = null;
-      let observable: ObservableQuery = null;
+      let observable: ObservableQuery<any> = null;
       const testQuery = gql`
         query {
           author {
@@ -310,7 +310,7 @@ describe('ObservableQuery', () => {
 
     it('does a network request if noFetch becomes false', (done) => {
       let queryManager: QueryManager = null;
-      let observable: ObservableQuery = null;
+      let observable: ObservableQuery<any> = null;
       const testQuery = gql`
         query {
           author {
@@ -355,7 +355,7 @@ describe('ObservableQuery', () => {
 
   describe('setVariables', () => {
     it('reruns query if the variables change', (done) => {
-      const observable: ObservableQuery = mockWatchQuery({
+      const observable: ObservableQuery<any> = mockWatchQuery({
         request: { query, variables },
         result: { data: dataOne },
       }, {
@@ -449,7 +449,7 @@ describe('ObservableQuery', () => {
     });
 
     it('reruns observer callback if the variables change but data does not', (done) => {
-      const observable: ObservableQuery = mockWatchQuery({
+      const observable: ObservableQuery<any> = mockWatchQuery({
         request: { query, variables },
         result: { data: dataOne },
       }, {
@@ -482,7 +482,7 @@ describe('ObservableQuery', () => {
 
       manager.query({ query, variables: differentVariables })
         .then(() => {
-          const observable: ObservableQuery = manager.watchQuery({ query, variables });
+          const observable: ObservableQuery<any> = manager.watchQuery({ query, variables });
 
           let errored = false;
           subscribeAndCount(done, observable, (handleCount, result) => {
@@ -501,7 +501,7 @@ describe('ObservableQuery', () => {
     });
 
     it('does not rerun query if variables do not change', (done) => {
-      const observable: ObservableQuery = mockWatchQuery({
+      const observable: ObservableQuery<any> = mockWatchQuery({
         request: { query, variables },
         result: { data: dataOne },
       }, {
@@ -528,7 +528,7 @@ describe('ObservableQuery', () => {
       // The expected behavior is that the original variables are forgotten
       // and the query stays in loading state until the result for the new variables
       // has returned.
-      const observable: ObservableQuery = mockWatchQuery({
+      const observable: ObservableQuery<any> = mockWatchQuery({
         request: { query, variables },
         result: { data: dataOne },
         delay: 20,
@@ -553,7 +553,7 @@ describe('ObservableQuery', () => {
 
   describe('currentResult', () => {
     it('returns the current query status immediately', (done) => {
-      const observable: ObservableQuery = mockWatchQuery({
+      const observable: ObservableQuery<any> = mockWatchQuery({
         request: { query, variables },
         result: { data: dataOne },
         delay: 100,