+++ test/QueryManager.ts
@@ -118,7 +118,7 @@ describe('QueryManager', () => {
     error?: Error,
     result?: ExecutionResult,
     delay?: number,
-    observer: Observer<ApolloQueryResult>,
+    observer: Observer<ApolloQueryResult<any>>,
   }) => {
     const queryManager = mockQueryManager({
       request: { query, variables },
@@ -127,7 +127,7 @@ describe('QueryManager', () => {
       delay,
     });
     const finalOptions = assign({ query, variables }, queryOptions) as WatchQueryOptions;
-    return queryManager.watchQuery(finalOptions).subscribe({
+    return queryManager.watchQuery<any>(finalOptions).subscribe({
       next: wrap(done, observer.next),
       error: observer.error,
     });
@@ -578,9 +578,9 @@ describe('QueryManager', () => {
     let subOneCount = 0;
 
     // pre populate data to avoid contention
-    queryManager.query(request)
+    queryManager.query<any>(request)
       .then(() => {
-        const handle = queryManager.watchQuery(request);
+        const handle = queryManager.watchQuery<any>(request);
 
         const subOne = handle.subscribe({
           next(result) {
@@ -654,7 +654,7 @@ describe('QueryManager', () => {
       secondResult: { data: data2 },
     });
 
-    const observable = queryManager.watchQuery(request);
+    const observable = queryManager.watchQuery<any>(request);
     return observableToPromise({ observable },
       (result) => {
         assert.deepEqual(result.data, data1);
@@ -695,7 +695,7 @@ describe('QueryManager', () => {
       secondResult: { data: data2 },
     });
 
-    const observable = queryManager.watchQuery(request);
+    const observable = queryManager.watchQuery<any>(request);
     return observableToPromise({ observable },
       (result) => {
         assert.deepEqual(result.data, data1);
@@ -736,7 +736,7 @@ describe('QueryManager', () => {
       secondResult: { data: data2 },
     });
 
-    const handle = queryManager.watchQuery(request);
+    const handle = queryManager.watchQuery<any>(request);
     handle.subscribe({});
 
     return handle.refetch().then(
@@ -804,7 +804,7 @@ describe('QueryManager', () => {
       },
     );
 
-    const observable = queryManager.watchQuery({ query });
+    const observable = queryManager.watchQuery<any>({ query });
     return observableToPromise({ observable },
       (result) => {
         assert.deepEqual(result.data, data1);
@@ -864,7 +864,7 @@ describe('QueryManager', () => {
       },
     );
 
-    const observable = queryManager.watchQuery({ query });
+    const observable = queryManager.watchQuery<any>({ query });
     const originalOptions = assign({}, observable.options);
     return observableToPromise({ observable },
       (result) => {
@@ -923,7 +923,7 @@ describe('QueryManager', () => {
       },
     );
 
-    const observable = queryManager.watchQuery({
+    const observable = queryManager.watchQuery<any>({
       query,
       pollInterval: 200,
     });
@@ -984,7 +984,7 @@ describe('QueryManager', () => {
       },
     );
 
-    const observable = queryManager.watchQuery({
+    const observable = queryManager.watchQuery<any>({
       query,
       pollInterval: 30,
       notifyOnNetworkStatusChange: true,
@@ -1059,10 +1059,10 @@ describe('QueryManager', () => {
     );
 
     // First, prime the store so that query diffing removes the query
-    return queryManager.query({
+    return queryManager.query<any>({
       query: primeQuery,
     }).then(() => {
-      const handle = queryManager.watchQuery({
+      const handle = queryManager.watchQuery<any>({
         query: complexQuery,
         returnPartialData: true,
       });
@@ -1130,10 +1130,10 @@ describe('QueryManager', () => {
     );
 
     // First, prime the cache
-    return queryManager.query({
+    return queryManager.query<any>({
       query: primeQuery,
     }).then(() => {
-      const handle = queryManager.watchQuery({
+      const handle = queryManager.watchQuery<any>({
         query: complexQuery,
         noFetch: true,
       });
@@ -1458,8 +1458,8 @@ describe('QueryManager', () => {
       },
     );
 
-    const observable1 = queryManager.watchQuery({ query: query1 });
-    const observable2 = queryManager.watchQuery({ query: query2 });
+    const observable1 = queryManager.watchQuery<any>({ query: query1 });
+    const observable2 = queryManager.watchQuery<any>({ query: query2 });
 
     return Promise.all([
       observableToPromise({ observable: observable1 },
@@ -1516,11 +1516,11 @@ describe('QueryManager', () => {
       },
     );
 
-    const observable = queryManager.watchQuery({ query: query1 });
+    const observable = queryManager.watchQuery<any>({ query: query1 });
     return observableToPromise({ observable },
       (result) => {
         assert.deepEqual(result.data, data1);
-        queryManager.query({ query: query2 });
+        queryManager.query<any>({ query: query2 });
       },
       // 3 because the query init action for the second query causes a callback
       (result) => assert.deepEqual(result.data, {
@@ -1567,7 +1567,7 @@ describe('QueryManager', () => {
           result: { data: data2 },
         },
       );
-      const observable = queryManager.watchQuery({
+      const observable = queryManager.watchQuery<any>({
         query,
         variables,
         pollInterval: 50,
@@ -1577,6 +1577,7 @@ describe('QueryManager', () => {
         (result) => assert.deepEqual(result.data, data1),
         (result) => assert.deepEqual(result.data, data2),
       );
+
     });
 
     it('should let you handle multiple polled queries and unsubscribe from one of them', (done) => {
@@ -1786,7 +1787,7 @@ describe('QueryManager', () => {
         },
       );
 
-      const observable = queryManager.watchQuery({
+      const observable = queryManager.watchQuery<any>({
         query,
         variables,
         pollInterval: 50,
@@ -1844,7 +1845,7 @@ describe('QueryManager', () => {
         },
       );
 
-      const observable = queryManager.watchQuery({ query, variables });
+      const observable = queryManager.watchQuery<any>({ query, variables });
       observable.startPolling(50);
 
       return observableToPromise({ observable },
@@ -1888,7 +1889,7 @@ describe('QueryManager', () => {
           result: { data: data2 },
         },
       );
-      const observable = queryManager.watchQuery({
+      const observable = queryManager.watchQuery<any>({
         query,
         variables,
         pollInterval: 50,
@@ -1965,7 +1966,7 @@ describe('QueryManager', () => {
   it('warns if you forget the template literal tag', () => {
     const queryManager = mockQueryManager();
     assert.throws(() => {
-      queryManager.query({
+      queryManager.query<any>({
         // Bamboozle TypeScript into letting us do this
         query: 'string' as any as DocumentNode,
       });
@@ -1979,7 +1980,7 @@ describe('QueryManager', () => {
     }, /wrap the query string in a "gql" tag/);
 
     assert.throws(() => {
-      queryManager.watchQuery({
+      queryManager.watchQuery<any>({
         // Bamboozle TypeScript into letting us do this
         query: 'string' as any as DocumentNode,
       });
@@ -2129,7 +2130,7 @@ describe('QueryManager', () => {
         },
       };
       queryManager = createQueryManager({ networkInterface });
-      const observable = queryManager.watchQuery({ query });
+      const observable = queryManager.watchQuery<any>({ query });
 
       // wait just to make sure the observable doesn't fire again
       return observableToPromise({ observable, wait: 0 },
@@ -2141,7 +2142,7 @@ describe('QueryManager', () => {
 
     it('should not refetch toredown queries', (done) => {
       let queryManager: QueryManager = null;
-      let observable: ObservableQuery = null;
+      let observable: ObservableQuery<any> = null;
       const query = gql`
         query {
           author {
@@ -2213,7 +2214,7 @@ describe('QueryManager', () => {
         },
       };
       queryManager = createQueryManager({ networkInterface });
-      const observable = queryManager.watchQuery({ query });
+      const observable = queryManager.watchQuery<any>({ query });
 
       // wait to make sure store reset happened
       return observableToPromise({ observable, wait: 20 },
@@ -2261,7 +2262,8 @@ describe('QueryManager', () => {
           }
         }`;
       const queryManager = mockQueryManager();
-      const mockObservableQuery: ObservableQuery = {
+
+      const mockObservableQuery: ObservableQuery<any> = {
         refetch(variables: any): Promise<ExecutionResult> {
           done();
           return null;
@@ -2270,10 +2272,10 @@ describe('QueryManager', () => {
           query: query,
         },
         scheduler: queryManager.scheduler,
-      } as any as ObservableQuery;
+      } as any as ObservableQuery<any>;
 
       const queryId = 'super-fake-id';
-      queryManager.addObservableQuery(queryId, mockObservableQuery);
+      queryManager.addObservableQuery<any>(queryId, mockObservableQuery);
       queryManager.resetStore();
     });
 
@@ -2290,7 +2292,7 @@ describe('QueryManager', () => {
       options.noFetch = true;
       options.query = query;
       let refetchCount = 0;
-      const mockObservableQuery: ObservableQuery = {
+      const mockObservableQuery: ObservableQuery<any> = {
         refetch(variables: any): Promise<ExecutionResult> {
           refetchCount ++;
           done();
@@ -2298,10 +2300,10 @@ describe('QueryManager', () => {
         },
         options,
         queryManager: queryManager,
-      } as any as ObservableQuery;
+      } as any as ObservableQuery<any>;
 
       const queryId = 'super-fake-id';
-      queryManager.addObservableQuery(queryId, mockObservableQuery);
+      queryManager.addObservableQuery<any>(queryId, mockObservableQuery);
       queryManager.resetStore();
       setTimeout(() => {
         assert.equal(refetchCount, 0);
@@ -2335,7 +2337,7 @@ describe('QueryManager', () => {
       };
 
       queryManager = createQueryManager({ networkInterface });
-      queryManager.query({ query }).then((result) => {
+      queryManager.query<any>({ query }).then((result) => {
         done(new Error('query() gave results on a store reset'));
       }).catch((error) => {
         done();
@@ -2449,10 +2451,10 @@ describe('QueryManager', () => {
         error: new Error('Network error ocurred'),
       },
     );
-    queryManager.query({ query }).then((result) => {
+    queryManager.query<any>({ query }).then((result) => {
       assert.deepEqual(result.data, data);
 
-      queryManager.query({ query, forceFetch: true }).then(() => {
+      queryManager.query<any>({ query, forceFetch: true }).then(() => {
         done(new Error('Returned a result when it was not supposed to.'));
       }).catch((error) => {
         // make that the error thrown doesn't empty the state
@@ -2520,7 +2522,7 @@ describe('QueryManager', () => {
         error: new Error('Network error occurred.'),
       },
     );
-    const observable = queryManager.watchQuery({ query, pollInterval: 20 });
+    const observable = queryManager.watchQuery<any>({ query, pollInterval: 20 });
 
     return observableToPromise({
         observable,
@@ -2570,7 +2572,7 @@ describe('QueryManager', () => {
       },
     );
 
-    const observable = queryManager.watchQuery({ query });
+    const observable = queryManager.watchQuery<any>({ query });
     return Promise.all<any[] | void>([
       // we wait for a little bit to ensure the result of the second query
       // don't trigger another subscription event
@@ -2579,7 +2581,7 @@ describe('QueryManager', () => {
           assert.deepEqual(result.data, data);
         },
       ),
-      queryManager.query({ query }).then((result) => {
+      queryManager.query<any>({ query }).then((result) => {
         assert.deepEqual(result.data, data);
       }),
     ]);
@@ -2682,8 +2684,8 @@ describe('QueryManager', () => {
       store,
     });
 
-    const observable1 = queryManager.watchQuery({ query: query1 });
-    const observable2 = queryManager.watchQuery({ query: query2 });
+    const observable1 = queryManager.watchQuery<any>({ query: query1 });
+    const observable2 = queryManager.watchQuery<any>({ query: query2 });
 
     // I'm not sure the waiting 60 here really is required, but the test used to do it
     return Promise.all([
@@ -2753,8 +2755,8 @@ describe('QueryManager', () => {
       store,
     });
 
-    const observableWithId = queryManager.watchQuery({ query: queryWithId });
-    const observableWithoutId = queryManager.watchQuery({ query: queryWithoutId });
+    const observableWithId = queryManager.watchQuery<any>({ query: queryWithId });
+    const observableWithoutId = queryManager.watchQuery<any>({ query: queryWithoutId });
 
     // I'm not sure the waiting 60 here really is required, but the test used to do it
     return Promise.all([
@@ -2841,8 +2843,8 @@ describe('QueryManager', () => {
       store,
     });
 
-    const observableWithId = queryManager.watchQuery({ query: queryWithId });
-    const observableWithoutId = queryManager.watchQuery({ query: queryWithoutId });
+    const observableWithId = queryManager.watchQuery<any>({ query: queryWithId });
+    const observableWithoutId = queryManager.watchQuery<any>({ query: queryWithoutId });
 
     // I'm not sure the waiting 60 here really is required, but the test used to do it
     return Promise.all([
@@ -2904,8 +2906,8 @@ describe('QueryManager', () => {
         },
       );
 
-      return queryManager.query({ query: primeQuery }).then((primeResult) => {
-        const observable = queryManager.watchQuery({ query, returnPartialData: true });
+      return queryManager.query<any>({ query: primeQuery }).then((primeResult) => {
+        const observable = queryManager.watchQuery<any>({ query, returnPartialData: true });
 
         return observableToPromise({ observable },
           (result) => {
@@ -3012,7 +3014,7 @@ describe('QueryManager', () => {
           result: { data: mutationData },
         },
       );
-      const observable = queryManager.watchQuery({ query });
+      const observable = queryManager.watchQuery<any>({ query });
       return observableToPromise({ observable },
         (result) => {
           assert.deepEqual(result.data, data);
@@ -3069,7 +3071,7 @@ describe('QueryManager', () => {
           result: { data: mutationData },
         },
       );
-      const observable = queryManager.watchQuery({ query });
+      const observable = queryManager.watchQuery<any>({ query });
       return observableToPromise({ observable },
         (result) => {
           assert.deepEqual(result.data, data);
@@ -3131,7 +3133,7 @@ describe('QueryManager', () => {
         },
       );
 
-      const observable = queryManager.watchQuery({ query });
+      const observable = queryManager.watchQuery<any>({ query });
       return observableToPromise({ observable },
         (result) => {
           assert.deepEqual(result.data, data);
@@ -3181,7 +3183,7 @@ describe('QueryManager', () => {
     it('transforms query() results', () => {
       response = {data: {foo: 123}};
       return client.query({query: gql`{ foo }`})
-        .then((result: ApolloQueryResult) => {
+        .then((result: ApolloQueryResult<any>) => {
           assert.deepEqual(result.data, {foo: 123, transformCount: 1});
         });
     });
@@ -3222,7 +3224,7 @@ describe('QueryManager', () => {
     it('transforms mutate() results', () => {
       response = {data: {foo: 123}};
       return client.mutate({mutation: gql`mutation makeChanges { foo }`})
-        .then((result: ApolloQueryResult) => {
+        .then((result: ApolloQueryResult<any>) => {
           assert.deepEqual(result.data, {foo: 123, transformCount: 1});
         });
     });
@@ -3245,11 +3247,11 @@ describe('QueryManager', () => {
 
       client = new ApolloClient({
         networkInterface,
-        resultTransformer(result: ApolloQueryResult) {
+        resultTransformer(result: ApolloQueryResult<any>) {
           result.data.__proto__ = Model.prototype;
           return result;
         },
-        resultComparator(result1: ApolloQueryResult, result2: ApolloQueryResult) {
+        resultComparator(result1: ApolloQueryResult<any>, result2: ApolloQueryResult<any>) {
           // A real example would, say, deep compare the two while ignoring prototypes.
           const foo1 = result1 && result1.data && result1.data.foo;
           const foo2 = result2 && result2.data && result2.data.foo;
@@ -3308,7 +3310,7 @@ describe('QueryManager', () => {
 
     const queryManager = mockRefetch({ request, firstResult, secondResult });
 
-    const handle = queryManager.watchQuery(request);
+    const handle = queryManager.watchQuery<any>(request);
 
     handle.subscribe({
       error: () => { /* nothing */ },